+++ src/index.js
@@ -4,15 +4,12 @@ import { applyMiddleware, createStore } from 'redux';
 import { Provider } from 'react-redux';
 import logger from 'redux-logger';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import allReducers from './reducers/reducer';
+import allReducers from './reducer';
 import AppContainer from './components/App/AppContainer';
 import registerServiceWorker from './registerServiceWorker';
 import './index.css';
 
-const store = createStore(
-  allReducers,
-  composeWithDevTools(applyMiddleware(logger)),
-);
+const store = createStore(allReducers, composeWithDevTools(applyMiddleware(logger)));
 
 ReactDOM.render(
   <Provider store={store}>
